requirements.txt
flask==1.0.2
flask-cors==3.0.6
Flask-SQLAlchemy==2.3.2
git+https://github.com/Manfed/PyLyrics
lxml==4.2.3
git+https://github.com/cmccandless/safepass